Add ExpenseService and ExpenseController
- Change money type from float/double to BigDecimal
- Add endpoint for getting total of incomes/expenses

+++ src/main/java/com/kuro/expensetracker/requests/CategoryRequest.java
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Builder
@@ -15,7 +16,7 @@ public class CategoryRequest {
     @NotNull
     private String name;
     private String description;
-    private Float threshold;
+    private BigDecimal threshold;
     private List<Transaction> transactions;
     private User owner;
 }
